Fraud script - Moved getAccessToken function to new file in common utils directory
