+++ deploy/scripts/src/cri-kiwi/BAV.ts
@@ -13,7 +13,8 @@ import { timeGroup } from '../common/utils/request/timing'
 import { isStatusCode200, pageContentCheck } from '../common/utils/checks/assertions'
 import { sleepBetween } from '../common/utils/sleep/sleepBetween'
 import { bankingPayload } from './data/BAVdata'
-import { getAuthorizeauthorizeLocation, getClientID, getCodeFromUrl, getAccessToken } from './utils/authorization'
+import { getAuthorizeauthorizeLocation, getClientID, getCodeFromUrl } from './utils/authorization'
+import { getAccessToken } from '../common/utils/authorization'
 import { getEnv } from '../common/utils/config/environment-variables'
 import { getThresholds } from '../common/utils/config/thresholds'
