+++ deploy/scripts/src/cri-kiwi/utils/authorization.ts
@@ -13,12 +13,6 @@ export function getCodeFromUrl(url: string): string {
   fail('Code not found')
 }
 
-export function getAccessToken(r: Response): string {
-  const accessToken = r.json('access_token')
-  if (accessToken !== null && typeof accessToken === 'string') return accessToken
-  fail('AccessToken not found')
-}
-
 export function getAuthorizeauthorizeLocation(r: Response): string {
   const authorizeLocation = r.json('AuthorizeLocation')
   if (authorizeLocation !== null && typeof authorizeLocation === 'string') return authorizeLocation
